webapi/tct-xmlhttprequest-w3c-tests/xmlhttprequest-py/xhr2_XMLHttpRequest_readyState_readonly.html

<!DOCTYPE html>
<!--
Copyright (c) 2014 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Liu, Shilong <shilongx.liu@intel.com>

-->

<html>
  <head>
    <title>XMLHttpRequest Test: xhr2_XMLHttpRequest_readyState_readonly</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://www.w3.org/TR/2012/WD-XMLHttpRequest-20121206/#dom-xmlhttprequest-readystate" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if XMLHttpRequest.readyState attribute is readonly" />
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
        var t = async_test(document.title);
        var client = new XMLHttpRequest();
        client.onreadystatechange = function () {
            t.step( function () {
                if (client.readyState == client.HEADERS_RECEIVED) {
                    var init_state = client.readyState;
                    client.readyState = client.UNSENT;
                    assert_equals(client.readyState, init_state, 'The XMLHttpRequest.readyState attribute is readonly.');
                    assert_not_equals(client.readyState, client.UNSENT ,"The XMLHttpRequest.readyState attribute should not be changed.")
                }
            });
            t.done();
        }
        client.open("GET", "support/test.xml");
        client.send();
    </script>
  </body>
</html>
